+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Fletch Changelog
 
 ## 0.2.0
-_Not Yet Released_
-- Added `.html` for getting/setting HTML content
-- Event listeners can now be removed
+_Released 7/20/2014_
+
+ * Added `.html` for getting/setting HTML content
+ 
+ * Added `.value` for getting/setting form input values
+ 
+ * Added `.checked` for getting/setting checked state for checkboxes and radio buttons
+ 
+ * Added `.serialize()` for encoding forms into URL parameters
+ 
+ * Event listeners can now be removed
 
 ## 0.1.0
 _Released 7/15/2014_
-- First release!
+
+ * First release!